requirements.txt
aiofiles==0.3.1
aiohttp==2.2.5
asn1crypto==0.22.0
async-timeout==1.2.1
backoff==1.4.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
flake8==3.4.1
httptools==0.0.9
idna==2.6
logzero==1.3.0
mccabe==0.6.1
motor==1.1
multidict==3.1.3
nexmo==1.5.0
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
PyJWT==1.5.2
pymongo==3.5.0
requests==2.18.3
sanic==0.6.0
six==1.10.0
ujson==1.35
urllib3==1.22
uvloop==0.8.0
websockets==3.3
yarl==0.12.0
